check for distributed

+++ one-shot-tuning.py
@@ -435,7 +435,7 @@ def main(
                     
                 control = batch.get("control_video")
                 if control is not None:
-                    control= rearrange(control, "b f c h w -> b c f h w")
+                    control= rearrange(control, "b f c h w -> b c f h w").to(weight_dtype)
 
                 model_pred = unet(noisy_latents, timesteps, clip_id, encoder_hidden_states,control=control).sample
                 ################
